#78 #79 [Refactor][fix] 실행 오류 해결
- ProjectDashboard, CommonCode package 경로 수정
- Role package 이름 수정 (Role -> role)
