--- src/main/java/org/omoknoone/ppm/commoncode/dto/CommonCodeGroupResponseDTO.java
@@ -1,14 +1,0 @@
-package org.omoknoone.ppm.commoncode.dto;
-
-import lombok.*;
-
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter
-@ToString
-public class CommonCodeGroupResponseDTO {
-
-    private Long groupId;
-    private String groupName;
-    private String groupDescription;
-}
