--- src/main/java/org/omoknoone/ppm/commoncode/dto/CommonCodeResponseDTO.java
@@ -1,14 +1,0 @@
-package org.omoknoone.ppm.commoncode.dto;
-
-import lombok.*;
-
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter
-@ToString
-public class CommonCodeResponseDTO {
-
-    private Long codeId;
-    private String codeName;
-    private String codeDescription;
-}
