--- src/main/java/org/omoknoone/ppm/commoncode/repository/CommonCodeGroupRepository.java
@@ -1,7 +1,0 @@
-package org.omoknoone.ppm.commoncode.repository;
-
-import org.omoknoone.ppm.commoncode.aggregate.CommonCodeGroup;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CommonCodeGroupRepository extends JpaRepository<CommonCodeGroup, Long> {
-}
