--- src/main/java/org/omoknoone/ppm/domain/Role/dto/ViewRoleListResponseDTO.java
@@ -1,15 +1,0 @@
-package org.omoknoone.ppm.domain.Role.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@AllArgsConstructor
-@NoArgsConstructor
-@Getter @Setter
-public class ViewRoleListResponseDTO {
-
-    private Integer roleId;
-    private String roleName;
-}
