--- src/main/java/org/omoknoone/ppm/domain/Role/repository/RoleRepository.java
@@ -1,7 +1,0 @@
-package org.omoknoone.ppm.domain.Role.repository;
-
-import org.omoknoone.ppm.domain.Role.aggregate.Role;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface RoleRepository extends JpaRepository<Role, Integer> {
-}
