+++ src/main/java/org/omoknoone/ppm/domain/commoncode/dto/CommonCodeGroupResponseDTO.java
@@ -1,0 +1,14 @@
+package org.omoknoone.ppm.domain.commoncode.dto;
+
+import lombok.*;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+@ToString
+public class CommonCodeGroupResponseDTO {
+
+    private Long groupId;
+    private String groupName;
+    private String groupDescription;
+}
