+++ src/main/java/org/omoknoone/ppm/domain/commoncode/dto/CommonCodeResponseDTO.java
@@ -1,0 +1,14 @@
+package org.omoknoone.ppm.domain.commoncode.dto;
+
+import lombok.*;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+@ToString
+public class CommonCodeResponseDTO {
+
+    private Long codeId;
+    private String codeName;
+    private String codeDescription;
+}
