+++ src/main/java/org/omoknoone/ppm/domain/commoncode/repository/CommonCodeGroupRepository.java
@@ -1,0 +1,7 @@
+package org.omoknoone.ppm.domain.commoncode.repository;
+
+import org.omoknoone.ppm.domain.commoncode.aggregate.CommonCodeGroup;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CommonCodeGroupRepository extends JpaRepository<CommonCodeGroup, Long> {
+}
