+++ src/main/java/org/omoknoone/ppm/domain/commoncode/service/CommonCodeService.java
@@ -1,0 +1,17 @@
+package org.omoknoone.ppm.domain.commoncode.service;
+
+import org.omoknoone.ppm.domain.commoncode.dto.CommonCodeResponseDTO;
+import org.omoknoone.ppm.domain.commoncode.dto.CommonCodeGroupResponseDTO;
+
+import java.util.List;
+
+public interface CommonCodeService {
+
+     List<CommonCodeResponseDTO> viewCommonCodesByGroup(Long groupId);
+
+     CommonCodeResponseDTO viewCommonCodeById(Long codeId);
+
+     List<CommonCodeGroupResponseDTO> viewAllCommonCodeGroups();
+
+     CommonCodeGroupResponseDTO viewCommonCodeGroupById(Long groupId);
+}
