+++ src/main/java/org/omoknoone/ppm/domain/projectDashboard/dto/ProjectDashboardDTO.java
@@ -1,0 +1,4 @@
+package org.omoknoone.ppm.domain.projectDashboard.dto;
+
+public class ProjectDashboardDTO {
+}
