+++ src/main/java/org/omoknoone/ppm/domain/projectDashboard/vo/ProjectDashboardVO.java
@@ -1,0 +1,4 @@
+package org.omoknoone.ppm.domain.projectDashboard.vo;
+
+public class ProjectDashboardVO {
+}
