+++ src/main/java/org/omoknoone/ppm/domain/role/service/RoleService.java
@@ -1,0 +1,9 @@
+package org.omoknoone.ppm.domain.role.service;
+
+import org.omoknoone.ppm.domain.role.dto.ViewRoleListResponseDTO;
+
+import java.util.List;
+
+public interface RoleService {
+    List<ViewRoleListResponseDTO> viewRoleList();
+}
