--- src/main/java/org/omoknoone/ppm/projectDashboard/dto/ProjectDashboardDTO.java
@@ -1,4 +1,0 @@
-package org.omoknoone.ppm.projectDashboard.dto;
-
-public class ProjectDashboardDTO {
-}
