--- src/main/java/org/omoknoone/ppm/projectDashboard/service/ProjectDashboardService.java
@@ -1,11 +1,0 @@
-package org.omoknoone.ppm.projectDashboard.service;
-
-import java.util.List;
-
-import org.omoknoone.ppm.projectDashboard.aggregate.ProjectDashboard;
-
-public interface ProjectDashboardService {
-	List<ProjectDashboard> viewProjectDashboardByProjectId(String projectId);
-	void updateGauge(String projectId);
-
-}
