--- src/main/java/org/omoknoone/ppm/projectDashboard/vo/ProjectDashboardVO.java
@@ -1,4 +1,0 @@
-package org.omoknoone.ppm.projectDashboard.vo;
-
-public class ProjectDashboardVO {
-}
